main refactoring
