+++ app/src/main/java/ru/learning/action/Action.java
@@ -1,0 +1,11 @@
+package ru.learning.action;
+
+public abstract class Action {
+    protected String actionText;
+
+    public void printlnAction(Integer number) {
+        System.out.println(actionText + " " + number + ": ");
+    }
+
+    public abstract Boolean action();
+}
