+++ app/src/main/java/ru/learning/action/ExitAction.java
@@ -1,0 +1,12 @@
+package ru.learning.action;
+
+public class ExitAction extends Action {
+    public ExitAction() {
+        actionText = "Для выхода из игры введите";
+    }
+
+    public Boolean action() {
+        System.out.println("Поверь, испытывать страх - это нормально. Когда ты это осознаешь, то сможешь ему сопротивляться.");
+        return true;
+    }
+}
